Intraceptor added, base url taken from system property in runtime
