+++ core/src/main/kotlin/io/specmatic/core/utilities/Flags.kt
@@ -11,7 +11,7 @@ class Flags {
         const val SPECMATIC_TEST_PARALLELISM = "SPECMATIC_TEST_PARALLELISM"
         const val SPECMATIC_STUB_DELAY = "SPECMATIC_STUB_DELAY"
         const val SPECMATIC_TEST_TIMEOUT = "SPECMATIC_TEST_TIMEOUT"
-
+        const val BASE_URL = "BASE_URL"
         const val CONFIG_FILE_PATH = "CONFIG_FILE_PATH"
 
         const val IGNORE_INLINE_EXAMPLES = "IGNORE_INLINE_EXAMPLES"
