+++ core/src/test/kotlin/io/specmatic/proxy/ProxyTest.kt
@@ -97,7 +97,7 @@ internal class ProxyTest {
     @Test
     fun `basic test of the proxy with a request containing a path variable`() {
         HttpStub(simpleFeature).use {
-            Proxy(host = "localhost", port = 9001, "", fakeFileWriter).use {
+            Proxy(host = "localhost", port = 9001, "http://localhost:9000", fakeFileWriter).use {
                 val restProxy = java.net.Proxy(java.net.Proxy.Type.HTTP, InetSocketAddress("localhost", 9001))
                 val requestFactory = SimpleClientHttpRequestFactory()
                 requestFactory.setProxy(restProxy)
